Se corrigen datos en la tabla y se modifica app.js para cargue de archivos

+++ migrations/20240517012940-create-product.js
@@ -15,11 +15,11 @@ module.exports = {
       },
       price_product: {
         allowNull: false,
-        type: Sequelize.DECIMAL
+        type: Sequelize.DECIMAL(10,2)
       },
       quantity_product: {
         allowNull: false,
-        type: Sequelize.DECIMAL
+        type: Sequelize.INTEGER
       },
       description: {
         allowNull: false,
